Gmail: Set a login hint of email address for authentication
Our becfm.ini file may specify the Gmail address to use for Create Centre roof data.  We can
pass this when authenticating to try to help users log in to the correct account.

My experiments seem to show this won't change the account if there is a default - the
default would need to be cleared first.