Filton weather: Alter all dates/times to GMT/UTC
We're aiming to simplify our database by storing dates and times all in GMT (so DateTimes
are always unique and we can use them as a primary key in our tables).

The dates and times from the Filton weather station are in local time.  We use our
dateTimeToGMT() function to change them to GMT on import.

In the process we move dateTimeToGMT out of the becsimtricity file to a new file called
miscfuncs.php...and move expandHomeDirectory() in there too.