Raspberry Pi Displays: Ensure iceweasel/firefox has started before pressing F11 to go full-screen
A simple sleep delay of 15 seconds might not always be sufficient for iceweasel/firefox to
start and display a browser window.  Instead we use xdotool to poll for the browser window
being opened.  We poll for both iceweasel and firefox as the the 1st SD card used iceweasel,
but the newer SD card downloaded firefox with an iceweasel wrapper, so it looks like
Raspian is moving towards using firefox.