Raspberry Pi Displays: Switch from using DPMS to tvservice for screen off/on
DPMS appeared not to work correctly for turning off displays in the crontab, but the tvservice
script appears to work correctly on my HDMI TV.  However, on my TV it was unable to
turn the screen back on if it had powered itself down.  The screen may need to be manually
turned back on in our community buildings if this is the case with their displays too.