Raspberry Pi Displays: When browser is full-screen, move the mouse pointer before hiding it
With the mouse pointer in the centre of the screen, as it is by default, the mouse-over
tool-tip of a BEC generation bar chart can be activated.  This does not disappear when
unclutter is used to hide the mouse pointer, so we move the mouse pointer out of the way
before hiding it.