HTML Slideshow: Support string replacement in URLs
In particular we support replacement of %SITECODE% with the site short code of the requested
site so that it is easy to have a common set of pages set up in the database which all sites
will be shown, but which are particular to their site (e.g. we only have to add a URL for the
chart of energy generation in the last 7 days once to the 'all' pseudo-site and all sites will
be shown their generation for the last 7 days.