Slideshow: Check slideshow URL can be fetched before attempting hourly reload
The slideshow page attempts to reload itself hourly.  If the URL cannot be loaded due to
the network/Internet being unavailable, an error page/blank page would be displayed and
the slideshow would be lost.

We use javascript to check the page can be loaded before attempting a full reload to prevent
connectivity issues breaking the slideshow.