Slideshow: Introduce an all_no_lim pseudo site
We add an all_no_lim pseudo site in addition to the all pseudo-site.  'all' is used for all sites,
but if we have data limits at a particular site (e.g. it's using a mobile data plan for Internet
access) then all_no_lim is excluded, otherwise it is included.

A column called data_limits was added to the slideshow_sites table to make this work.