Raspberry Pi Displays: Minor alterations
Coniston want to switch from 8am to 10pm to 9am to 5pm as this is when the cafe area is
generally open.  A way is needed to make the on/off times of the screens site specific
(or start using a PIR sensor before too much longer).

Also alter where the mouse is moved to before hiding the pointer as (1, 1), in the top-left
corner, now causes the Firefox menu bar to appear - we move it down a little.