Raspberry Pi Displays: becdisplayonly: The scheduled update script for a low-data Raspberry Pi display
This script should run before the slide-show is loaded and then at maybe 9am each day.  It loads
power/reading data into the local copy of the BEC database; the local copy just has a few of
the tables from the livegen website: enough to serve up the slide-show, charts and headlines
pages locally.

NOTE: At time of addition this is only used to seed the bec MySQL database on the Pi - there
is still work to do to add a service to get graphing data from the local database.