Raspberry Pi Displays: make_low_data_pi_www: Script to create web side of a low-data Raspberry Pi Display
This builds the graphing website, brings a little extra data local to the Pi and merges in the
bec_services repository.

It is a little more interactive than the script for a full-data Raspberry Pi display.  The oldest
version of AngularJS should be picked when prompted.

The script will copy the web site over to the Pi if given the PI_IP_ADDR.

It will also leave a copy in /tmp/bec_pi/www which could be tested locally.

At the moment the generation data still comes from Simtricity on each page refresh, but the
idea is to enhance it so that the data is stored in the local MySQL database and a new service
in the graphing website will extract it from there.  Then there need only be one fetch of data
per day, regardless of the number of times the Pi is turned on/off.