Raspberry Pi Displays: Install script for a low-data Raspberry Pi display
This is a variant of the raspberry_install.sh script.  It installs the extra
packages to server up the slide-show and data web-site from the Pi rather than
retrieving it from livegen.bristolenergy.coop.

In addition to the usual steps Apache and MySQL are configured.

Lots of Firefox features are disabled to prevent unrequested data transfer.

The NTP service will be stopped after the first synchronisation.

The avahi-daemon (bonjour service) will be stopped.

We remove the crontab initially as the first user is Wick Sports Ground where the
Pi will be shared with a TV and turned on/off with the TV rather than needing to
turn the screen on and off on a schedule.