Simtricity: Ignore meters without site names
A few days ago another meter crept into the list returned by Simtricity that has not yet been
assigned a site name.  This was causing the script to die with an error.

This patch changes the becsimtricity.php script to ignore meters which don't have a site
name in them.  We still print an error so it can be seen and reported to CEPro/Simtricity.